fix: resolve CI code style and lint violations
- Replace .count == 0 with .isEmpty for better Swift practices
- Fix long lines exceeding 200 characters by splitting expressions
- Refactor functions with >5 parameters using parameter structs
  - Created PDFGeneratorParams.swift with parameter structs
  - Updated all test helpers to use parameter structs
- Split large files (>300 lines) into smaller modules using extensions
  - Split test files into extension files for better organization
  - All files now comply with ≤300 line limit

All changes verified with successful build and test build.
