+++ PayslipMaxTests/Helpers/BasicPayslipGenerator.swift
@@ -4,47 +4,14 @@ import PDFKit
 
 /// Protocol for basic payslip generation operations
 protocol BasicPayslipGeneratorProtocol {
-    /// Creates a standard military payslip for testing
-    static func standardMilitaryPayslip(
-        id: UUID,
-        month: String,
-        year: Int,
-        rank: String,
-        name: String,
-        serviceNumber: String,
-        credits: Double,
-        debits: Double,
-        dsop: Double,
-        tax: Double,
-        includeAllowances: Bool
-    ) -> PayslipItem
+    /// Creates a standard military payslip for testing using parameter struct
+    static func standardMilitaryPayslip(params: MilitaryPayslipParams) -> PayslipItem
 
-    /// Creates a PCDA payslip for testing defense personnel
-    static func standardPCDAPayslip(
-        id: UUID,
-        month: String,
-        year: Int,
-        name: String,
-        serviceNumber: String,
-        basicPay: Double,
-        msp: Double,
-        da: Double,
-        dsop: Double,
-        incomeTax: Double
-    ) -> PayslipItem
+    /// Creates a PCDA payslip for testing defense personnel using parameter struct
+    static func standardPCDAPayslip(params: PCDAPayslipParams) -> PayslipItem
 
-    /// Creates a military payslip PDF for testing
-    static func militaryPayslipPDF(
-        name: String,
-        rank: String,
-        serviceNumber: String,
-        month: String,
-        year: Int,
-        credits: Double,
-        debits: Double,
-        dsop: Double,
-        tax: Double
-    ) -> PDFDocument
+    /// Creates a military payslip PDF for testing using parameter struct
+    static func militaryPayslipPDF(params: PayslipPDFParams) -> PDFDocument
 }
 
 /// A generator for basic payslip-related test data
@@ -52,29 +19,17 @@ class BasicPayslipGenerator: BasicPayslipGeneratorProtocol {
 
     // MARK: - Standard Payslip Data Generation
 
-    /// Creates a standard military payslip for testing
-    static func standardMilitaryPayslip(
-        id: UUID = UUID(),
-        month: String = "January",
-        year: Int = 2023,
-        rank: String = "Major",
-        name: String = "John Doe",
-        serviceNumber: String = "MIL123456",
-        credits: Double = 85000.0,
-        debits: Double = 15000.0,
-        dsop: Double = 6000.0,
-        tax: Double = 25000.0,
-        includeAllowances: Bool = true
-    ) -> PayslipItem {
+    /// Creates a standard military payslip for testing using parameter struct
+    static func standardMilitaryPayslip(params: MilitaryPayslipParams = .default) -> PayslipItem {
         let payslip = PayslipItem(
-            id: id,
-            month: month,
-            year: year,
-            credits: credits,
-            debits: debits,
-            dsop: dsop,
-            tax: tax,
-            name: name,
+            id: params.id,
+            month: params.month,
+            year: params.year,
+            credits: params.credits,
+            debits: params.debits,
+            dsop: params.dsop,
+            tax: params.tax,
+            name: params.name,
             accountNumber: "XXXX5678",
             panNumber: "ABCDE1234F"
         )
@@ -85,33 +40,17 @@ class BasicPayslipGenerator: BasicPayslipGeneratorProtocol {
         return payslip
     }
 
-    /// Creates a PCDA payslip for testing defense personnel
-    static func standardPCDAPayslip(
-        id: UUID = UUID(),
-        month: String = "January",
-        year: Int = 2023,
-        name: String = "Major Jane Smith",
-        serviceNumber: String = "PCDA123456",
-        basicPay: Double = 67700.0,
-        msp: Double = 15500.0,
-        da: Double = 4062.0,
-        dsop: Double = 6770.0,
-        incomeTax: Double = 15000.0
-    ) -> PayslipItem {
-        // PCDA payslip fields
-        let credits = basicPay + msp + da
-        let debits = dsop
-        let tax = incomeTax
-
+    /// Creates a PCDA payslip for testing defense personnel using parameter struct
+    static func standardPCDAPayslip(params: PCDAPayslipParams = .default) -> PayslipItem {
         let payslip = PayslipItem(
-            id: id,
-            month: month,
-            year: year,
-            credits: credits,
-            debits: debits,
-            dsop: dsop,  // PCDA payslips include DSOP
-            tax: tax,
-            name: name,
+            id: params.id,
+            month: params.month,
+            year: params.year,
+            credits: params.credits,
+            debits: params.dsop,
+            dsop: params.dsop,  // PCDA payslips include DSOP
+            tax: params.incomeTax,
+            name: params.name,
             accountNumber: "XXXX4321",
             panNumber: "FGHIJ5678K"
         )
@@ -123,33 +62,24 @@ class BasicPayslipGenerator: BasicPayslipGeneratorProtocol {
 
     // MARK: - PDF Generation
 
-    /// Creates a military payslip PDF for testing
-    static func militaryPayslipPDF(
-        name: String = "John Doe",
-        rank: String = "Major",
-        serviceNumber: String = "MIL123456",
-        month: String = "January",
-        year: Int = 2023,
-        credits: Double = 85000.0,
-        debits: Double = 15000.0,
-        dsop: Double = 6000.0,
-        tax: Double = 25000.0
-    ) -> PDFDocument {
+    /// Creates a military payslip PDF for testing using parameter struct
+    static func militaryPayslipPDF(params: PayslipPDFParams = .default) -> PDFDocument {
         // Utilize the core PDF generation capability from the main test data generator
-        return TestDataGenerator.samplePayslipPDF(
-            serviceBranch: .army, // Default to army
-            name: name,
-            rank: rank,
-            serviceNumber: serviceNumber,
-            month: month,
-            year: year,
-            basicPay: credits,
-            msp: debits,
-            da: 5610.0, // Default Dearness Allowance
-            dsop: dsop,
-            agif: 150.0, // Default AGIF
-            incomeTax: tax
+        let defenseParams = DefensePayslipPDFParams(
+            serviceBranch: .army,
+            name: params.name,
+            rank: params.rank,
+            serviceNumber: params.id,
+            month: params.month,
+            year: params.year,
+            basicPay: params.credits,
+            msp: params.debits,
+            da: 5610.0,
+            dsop: params.dsop,
+            agif: 150.0,
+            incomeTax: params.tax
         )
+        return TestDataGenerator.samplePayslipPDF(params: defenseParams)
     }
 
     // MARK: - Helper Methods
