+++ PayslipMaxTests/Helpers/ScenarioBuilder.swift
@@ -48,7 +48,7 @@ class ScenarioBuilder: ScenarioBuilderProtocol {
     // MARK: - ScenarioBuilderProtocol Implementation
 
     func buildMonthlyScenario(for month: String, year: Int, baseAmount: Double) -> TestScenario {
-        let payslip = dataFactory.createPayslipItem(
+        let params = PayslipItemParams(
             id: UUID(),
             month: month,
             year: year,
@@ -60,6 +60,7 @@ class ScenarioBuilder: ScenarioBuilderProtocol {
             accountNumber: "TEST123456",
             panNumber: "TESTP1234F"
         )
+        let payslip = dataFactory.createPayslipItem(params: params)
 
         let expectedTotalCredits = baseAmount
         let expectedTotalDebits = baseAmount * 0.2 + baseAmount * 0.05 + baseAmount * 0.15
@@ -84,7 +85,7 @@ class ScenarioBuilder: ScenarioBuilderProtocol {
             let yearOffset = index / 12
             let monthlyVariation = Double.random(in: 0.8...1.2) // ±20% variation
 
-            return dataFactory.createPayslipItem(
+            let params = PayslipItemParams(
                 id: UUID(),
                 month: month,
                 year: startingYear + yearOffset,
@@ -96,6 +97,7 @@ class ScenarioBuilder: ScenarioBuilderProtocol {
                 accountNumber: "YEARLY123456",
                 panNumber: "YEART1234F"
             )
+            return dataFactory.createPayslipItem(params: params)
         }
 
         let totalCredits = payslips.reduce(0) { $0 + $1.credits }
@@ -153,15 +155,12 @@ class ScenarioBuilder: ScenarioBuilderProtocol {
     // MARK: - Private Helper Methods
 
     private func buildZeroBalanceScenario() -> TestScenario {
-        let payslip = dataFactory.createPayslipItem(
-            id: UUID(),
-            month: "January",
-            year: 2023,
+        let params = PayslipItemParams(
+            id: UUID(), month: "January", year: 2023,
             credits: 0, debits: 0, dsop: 0, tax: 0,
-            name: "Zero Balance Employee",
-            accountNumber: "ZERO000000",
-            panNumber: "ZEROZ1234F"
+            name: "Zero Balance Employee", accountNumber: "ZERO000000", panNumber: "ZEROZ1234F"
         )
+        let payslip = dataFactory.createPayslipItem(params: params)
 
         return TestScenario(
             title: "Zero Balance Edge Case",
@@ -176,18 +175,12 @@ class ScenarioBuilder: ScenarioBuilderProtocol {
 
     private func buildHighValueScenario() -> TestScenario {
         let highAmount = 1_000_000.0
-        let payslip = dataFactory.createPayslipItem(
-            id: UUID(),
-            month: "January",
-            year: 2023,
-            credits: highAmount,
-            debits: highAmount * 0.3,
-            dsop: highAmount * 0.05,
-            tax: highAmount * 0.2,
-            name: "High Value Employee",
-            accountNumber: "HIGH123456",
-            panNumber: "HIGHP1234F"
+        let params = PayslipItemParams(
+            id: UUID(), month: "January", year: 2023,
+            credits: highAmount, debits: highAmount * 0.3, dsop: highAmount * 0.05, tax: highAmount * 0.2,
+            name: "High Value Employee", accountNumber: "HIGH123456", panNumber: "HIGHP1234F"
         )
+        let payslip = dataFactory.createPayslipItem(params: params)
 
         let expectedTotalDebits = highAmount * 0.3 + highAmount * 0.05 + highAmount * 0.2
 
@@ -203,18 +196,12 @@ class ScenarioBuilder: ScenarioBuilderProtocol {
     }
 
     private func buildNegativeBalanceScenario() -> TestScenario {
-        let payslip = dataFactory.createPayslipItem(
-            id: UUID(),
-            month: "January",
-            year: 2023,
-            credits: 1000,
-            debits: 1500,
-            dsop: 300,
-            tax: 200,
-            name: "Negative Balance Employee",
-            accountNumber: "NEG123456",
-            panNumber: "NEGAT1234F"
+        let params = PayslipItemParams(
+            id: UUID(), month: "January", year: 2023,
+            credits: 1000, debits: 1500, dsop: 300, tax: 200,
+            name: "Negative Balance Employee", accountNumber: "NEG123456", panNumber: "NEGAT1234F"
         )
+        let payslip = dataFactory.createPayslipItem(params: params)
 
         let expectedTotalDebits = 1500 + 300 + 200
 
@@ -230,18 +217,12 @@ class ScenarioBuilder: ScenarioBuilderProtocol {
     }
 
     private func buildSpecialCharactersScenario() -> TestScenario {
-        let payslip = dataFactory.createPayslipItem(
-            id: UUID(),
-            month: "January",
-            year: 2023,
-            credits: 5000.0,
-            debits: 1000.0,
-            dsop: 300.0,
-            tax: 800.0,
-            name: "José María O'Connor-Smith, Jr.",
-            accountNumber: "TEST-123/456.789",
-            panNumber: "TESTP1234F&"
+        let params = PayslipItemParams(
+            id: UUID(), month: "January", year: 2023,
+            credits: 5000.0, debits: 1000.0, dsop: 300.0, tax: 800.0,
+            name: "José María O'Connor-Smith, Jr.", accountNumber: "TEST-123/456.789", panNumber: "TESTP1234F&"
         )
+        let payslip = dataFactory.createPayslipItem(params: params)
 
         return TestScenario(
             title: "Special Characters Edge Case",
@@ -256,18 +237,15 @@ class ScenarioBuilder: ScenarioBuilderProtocol {
 
     private func buildLargeDataSetScenario() -> TestScenario {
         let largePayslips = (1...50).map { index in
-            dataFactory.createPayslipItem(
-                id: UUID(),
-                month: "January",
-                year: 2023,
-                credits: Double(1000 + index * 100),
-                debits: Double(200 + index * 10),
-                dsop: Double(50 + index * 5),
-                tax: Double(150 + index * 15),
+            let params = PayslipItemParams(
+                id: UUID(), month: "January", year: 2023,
+                credits: Double(1000 + index * 100), debits: Double(200 + index * 10),
+                dsop: Double(50 + index * 5), tax: Double(150 + index * 15),
                 name: "Large Dataset Employee \(index)",
                 accountNumber: "LARGE\(String(format: "%04d", index))",
                 panNumber: "LARGE\(String(format: "%04d", index))F"
             )
+            return dataFactory.createPayslipItem(params: params)
         }
 
         let totalCredits = largePayslips.reduce(0) { $0 + $1.credits }
@@ -285,44 +263,22 @@ class ScenarioBuilder: ScenarioBuilderProtocol {
     }
 
     private func buildStandardVariations() -> [PayslipItem] {
-        return [
-            dataFactory.createPayslipItem(
-                id: UUID(),
-                month: "January",
-                year: 2023,
-                credits: 5000,
-                debits: 1000,
-                dsop: 300,
-                tax: 800,
-                name: "Standard Employee 1",
-                accountNumber: "STD001234",
-                panNumber: "STD001234F"
-            ),
-            dataFactory.createPayslipItem(
-                id: UUID(),
-                month: "January",
-                year: 2023,
-                credits: 6000,
-                debits: 1200,
-                dsop: 400,
-                tax: 900,
-                name: "Standard Employee 2",
-                accountNumber: "STD002345",
-                panNumber: "STD002345F"
-            ),
-            dataFactory.createPayslipItem(
-                id: UUID(),
-                month: "January",
-                year: 2023,
-                credits: 7000,
-                debits: 1400,
-                dsop: 500,
-                tax: 1000,
-                name: "Standard Employee 3",
-                accountNumber: "STD003456",
-                panNumber: "STD003456F"
-            )
+        let variations: [(credits: Double, debits: Double, dsop: Double, tax: Double, empNum: Int)] = [
+            (5000, 1000, 300, 800, 1),
+            (6000, 1200, 400, 900, 2),
+            (7000, 1400, 500, 1000, 3)
         ]
+
+        return variations.map { v in
+            let params = PayslipItemParams(
+                id: UUID(), month: "January", year: 2023,
+                credits: v.credits, debits: v.debits, dsop: v.dsop, tax: v.tax,
+                name: "Standard Employee \(v.empNum)",
+                accountNumber: "STD00\(v.empNum)234",
+                panNumber: "STD00\(v.empNum)234F"
+            )
+            return dataFactory.createPayslipItem(params: params)
+        }
     }
 
     private func buildEdgeCaseVariations() -> [PayslipItem] {
